Complete list UI and API for adding and deleting items

+++ home/src/components/DataItemList.js
@@ -10,7 +10,7 @@ import TableCell from '@material-ui/core/TableCell';
 import TableHead from '@material-ui/core/TableHead';
 import TableRow from '@material-ui/core/TableRow';
 import {getList} from '../apiClient';
-import PropTypes from 'prop-types'
+import PropTypes from 'prop-types';
 import Button from "@material-ui/core/Button";
 import DeleteIcon from "@material-ui/icons/Delete";
 import AddIcon from "@material-ui/icons/Add";
@@ -18,7 +18,6 @@ import {deleteItem, addItem} from '../apiClient';
 import Dialog from "@material-ui/core/Dialog";
 import DialogTitle from "@material-ui/core/DialogTitle";
 import DialogContent from "@material-ui/core/DialogContent";
-import DialogContentText from "@material-ui/core/DialogContentText";
 import TextField from "@material-ui/core/TextField";
 import DialogActions from "@material-ui/core/DialogActions";
 
@@ -31,21 +30,29 @@ class DataItemList extends React.Component {
   constructor(props) {
     super(props);
     this.mounted = true;
+    this.deleteConfirmTimeout = null;
+    this.dialog = {};
     this.state = {
       description: null,
       items: null,
       dialogIsOpen: false,
+      deleteConfirm: null,
       newItemName: '',
       newItemDescription: '',
     }
   }
 
   componentDidMount() {
-    getList(this.props.title).then(response => {
+    getList(this.props.title, true).then(response => {
       if(this.mounted) {
+        if(!response.body.length) {
+          //list doesn't exist
+          return
+        }
+        const {items, description} = response.body[0];
         this.setState({
-          description: response.description,
-          items: response.items,
+          description,
+          items,
         })
       }
     })
@@ -57,17 +64,20 @@ class DataItemList extends React.Component {
 
   closeDialog = (isConfirm) => {
     if(isConfirm) {
+      if(!this.state.newItemName) {
+        return;
+      }
       const item = {
         name: this.state.newItemName,
         description: this.state.newItemDescription,
       };
       addItem(this.props.title, item.name, item.description)
-      .catch(error => {
-        console.error(`could not create item (${item.name}): ${error}`);
-        this.state.items.pop()
-      });
+        .catch(error => {
+          console.error(`could not create item (${item.name}): ${error}`);
+          this.state.items.pop();
+          this.forceUpdate();
+        });
       this.state.items.push(item);
-      this.setState({items: this.state.items});
     }
     this.setState({
       dialogIsOpen: false,
@@ -79,42 +89,65 @@ class DataItemList extends React.Component {
   deleteRow = (index) => {
     const item = this.state.items[index];
     return () => {
+      const {deleteConfirm} = this.state;
+      if(this.deleteConfirmTimeout) {
+        clearTimeout(this.deleteConfirmTimeout);
+        this.deleteConfirmTimeout = undefined;
+      }
+      if(deleteConfirm !== index) {
+        this.deleteConfirmTimeout = setTimeout(() => this.setState({deleteConfirm: null}), 3000);
+        this.setState({deleteConfirm: index});
+        return;
+      }
       deleteItem(this.props.title, item.name, item.description)
         .catch(error => {
           // Print the error and return the item to the list.
           console.error(`could not delete item (${item.name}): ${error}`);
           this.state.items.splice(index, 0, item);
+          this.forceUpdate();
         });
       // Remove the item from the list.
       this.state.items.splice(index, 1);
-      this.setState({items: this.state.items})
+      this.setState({deleteConfirm: null})
     }
+  };
+
+  createRow = (row, index) => {
+    const {deleteConfirm} = this.state;
+    const {classes, theme} = this.props;
+
+    return <TableRow key={index} className={classes.row} style={{backgroundColor: index === deleteConfirm ? 'pink' : theme.palette.background.default}}>
+      <TableCell className={classes.cell} align='left'>
+        <Button
+          className={classes.deleteRowButton} size='small' onClick={this.deleteRow(index)}>
+          <DeleteIcon className={classes.deleteRowIcon}/>
+        </Button>
+      </TableCell>
+      <TableCell className={classes.cell} component="th" scope="row">{row.name}</TableCell>
+      <TableCell className={classes.cell} align="right">{row.description}</TableCell>
+    </TableRow>
   };
 
   render() {
     const {title, classes} = this.props;
     const {items, dialogIsOpen} = this.state;
-    console.log(this.props);
     return <Paper className={classes.root}>
-      <Dialog open={dialogIsOpen} onClose={() => this.closeDialog()} aria-labelledby="form-dialog-title">
+      <Dialog scroll='body' open={dialogIsOpen} onClose={() => this.closeDialog()}>
         <DialogTitle id="form-dialog-title">New Item</DialogTitle>
-        <DialogContent>
-          <DialogContentText>
-            Create a new item in the list with the following name and description.
-          </DialogContentText>
+        <DialogContent onSubmit={() => this.closeDialog(true)}>
           <TextField
             required
             autoFocus
             margin="dense"
             label="Item Name"
             fullWidth
+            onChange={(e) => (!this.state.newItemName || !e.target.value) && this.setState({newItemName: e.target.value})}
             onBlur={(e) => this.setState({newItemName: e.target.value})}
           />
           <TextField
             margin="dense"
             id="name"
             label="Item Description"
-            type="email"
             fullWidth
             onBlur={(e) => this.setState({newItemDescription: e.target.value})}
           />
@@ -124,48 +157,40 @@ class DataItemList extends React.Component {
             Cancel
           </Button>
           <Button disabled={!this.state.newItemName} onClick={() => this.closeDialog(true)} color="primary">
-            Subscribe
+            Add
           </Button>
         </DialogActions>
       </Dialog>
 
       <Typography variant='h3'>{title}</Typography>
-      <div className={''}>
-        <Button className={classes.createRowButton} size='small' onClick={() => this.setState({dialogIsOpen: true})}>
-          <AddIcon/>
-          <Typography>New Item</Typography>
-        </Button>
         {
-          items !== null && <Table className={classes.table}>
-            <TableHead>
-              <TableRow>
-                <TableCell>Action</TableCell>
-                <TableCell>Item</TableCell>
-                <TableCell align="right">Description</TableCell>
-              </TableRow>
-            </TableHead>
-            <TableBody className={classes.content}>
-              {items.map((row, index) => (
-                <TableRow key={index}>
-                  <TableCell className={classes.cell} align='left'>
-                    <Button className={classes.deleteRowButton} size='small' onClick={this.deleteRow(index)}>
-                      <DeleteIcon className={classes.deleteRowIcon}/>
-                    </Button>
-                  </TableCell>
-                  <TableCell className={classes.cell} component="th" scope="row">{row.name}</TableCell>
-                  <TableCell className={classes.cell} align="right">{row.description}</TableCell>
-                </TableRow>
-              ))}
-            </TableBody>
-          </Table>
+          items !== null
+            ?
+            <div>
+              <Button className={classes.createRowButton} size='small' onClick={() => this.setState({dialogIsOpen: true})}>
+                <AddIcon/>
+                <Typography>New Item</Typography>
+              </Button>
+
+              <Table className={classes.table}>
+                <TableHead>
+                  <TableRow>
+                    <TableCell>Action</TableCell>
+                    <TableCell>Item</TableCell>
+                    <TableCell align="right">Description</TableCell>
+                  </TableRow>
+                </TableHead>
+                <TableBody className={classes.content}>
+                  {items.map(this.createRow)}
+                </TableBody>
+              </Table>
+            </div>
+            :
+            <div>
+              <Typography>Loading...</Typography>
+              <CircularProgress variant='indeterminate' size={20}/>
+            </div>
         }
-        {
-          items === null && <div>
-            <Typography>Loading...</Typography>
-            <CircularProgress variant='indeterminate' size={20}/>
-          </div>
-        }
-      </div>
     </Paper>
   }
 }
@@ -192,10 +217,11 @@ export default withStyles(theme => ({
     borderTop: 'groove'
   },
   row: {
-    marginTop: 20,
-    border: 'solid',
-    borderSize: 1,
-    borderColor: 'black',
+    transition: 'backgroundColor 1s ease, color 1s ease'
+  },
+  errorBackground: {
+    backgroundColor: theme.palette.error.light,
+    color: theme.palette.error.contrastText,
   },
   cell: {
     padding: 1,
@@ -208,12 +234,13 @@ export default withStyles(theme => ({
     backgroundColor: theme.palette.error.main,
     color: theme.palette.error.contrastText,
     '&hover': {
+      opacity: 1,
       backgroundColor: theme.palette.error.dark,
       color: theme.palette.error.contrastText,
-    }
+    },
   },
   deleteRowIcon: {
     fontSize: 18,
     paddingRight: 2
   }
-}))(DataItemList)
+}), {withTheme: true})(DataItemList)
